Move the bukkit parsers into parser-bukkit
