+++ command-base/pom.xml
@@ -17,11 +17,13 @@
             <groupId>uk.co.thinkofdeath</groupId>
             <artifactId>parser-base</artifactId>
             <version>${project.version}</version>
+            <scope>compile</scope>
         </dependency>
         <dependency>
             <groupId>uk.co.thinkofdeath</groupId>
             <artifactId>common</artifactId>
             <version>${project.version}</version>
+            <scope>compile</scope>
         </dependency>
     </dependencies>
 </project>
