+++ parser-bukkit/pom.xml
@@ -25,7 +25,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>parser-base</artifactId>
+    <artifactId>parser-bukkit</artifactId>
     <version>0.2-SNAPSHOT</version>
 
     <dependencies>
@@ -33,7 +33,17 @@
             <groupId>uk.co.thinkofdeath</groupId>
             <artifactId>common</artifactId>
             <version>${project.version}</version>
-            <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>uk.co.thinkofdeath</groupId>
+            <artifactId>parser-base</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.spigotmc</groupId>
+            <artifactId>spigot-api</artifactId>
+            <version>1.7.10-R0.1-SNAPSHOT</version>
+            <scope>provided</scope>
         </dependency>
     </dependencies>
 </project>
