+++ parser-bukkit/src/main/java/uk/co/thinkfdeath/parsing/bukkit/parsing/PlayerParser.java
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-package uk.co.thinkofdeath.command.bukkit;
+package uk.co.thinkfdeath.parsing.bukkit.parsing;
 
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.Plugin;
-import uk.co.thinkofdeath.command.CommandError;
 import uk.co.thinkofdeath.parsing.parsers.ArgumentParser;
 import uk.co.thinkofdeath.parsing.ParserException;
 
@@ -28,7 +27,7 @@ import java.util.Set;
 /**
  * Parses player name's into Player objects. By default
  * this parsing doesn't require an exact match to work.
- * This may be changed by using {@link uk.co.thinkofdeath.command.bukkit.Strict}
+ * This may be changed by using {@link uk.co.thinkfdeath.parsing.bukkit.validators.Strict}
  */
 public class PlayerParser implements ArgumentParser<Player> {
     private final Plugin plugin;
